Support complex resolvers
