+++ .travis.yml
@@ -1,7 +1,6 @@
 language: node_js
 after_success: npm run coverage
 node_js:
-  - 10
   - 12
   - 14
   - node
