+++ package.json
@@ -3,6 +3,9 @@
   "version": "0.6.0",
   "description": "Simplify your schema by combining allOf into the root schema, safely.",
   "main": "src/index.js",
+  "engines": {
+    "node": ">=12.0.0"
+  },
   "scripts": {
     "eslint": "eslint src test",
     "test": "npm run eslint && nyc --reporter=html --reporter=text mocha test/specs",
